Zoltan2: Working Cuda C=22 K=8
Resolved a warning but also verified this checkpoint
is passing all Cuda tests and the problematic C=22 K=8
bug is also fixed.

+++ packages/zoltan2/src/algorithms/partition/Zoltan2_AlgMultiJagged.hpp
@@ -4597,7 +4597,7 @@ clock_weights_new_to_optimize.stop();
 
   clock_weights3.start();
 
-  mj_scalar_t * part_weights = new mj_scalar_t[array_length];
+  mj_scalar_t * part_weights = new mj_scalar_t[static_cast<size_t>(array_length)];
 
   ReduceWeightsFunctor<policy_t, mj_scalar_t, mj_part_t, mj_lno_t,
     typename mj_node_t::device_type>
